requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
werkzeug==2.0.1
yfinance==0.1.70
ib_insync==0.9.70
python-dotenv==1.0.1
pytz==2023.3
alpaca-trade-api==3.2.0
websockets==10.4

